fix(jules-cli): move get_api_key to config.py to avoid circular/inline imports
