+++ src/python/jules_cli/jules_cli/commands/activity.py
@@ -4,14 +4,13 @@ import json
 import typer
 
 from jules_cli.client import JulesClient
+from jules_cli.config import get_api_key
 from jules_cli.core import JulesContext
 
 activity_app = typer.Typer(add_completion=False, help="Manage activities.")
 
 
 def _get_api_key(ctx: typer.Context) -> str:
-    from jules_cli.main import get_api_key  # noqa: PLC0415
-
     jules_ctx: JulesContext = ctx.obj
     return get_api_key(jules_ctx.api_key)
